Track page changes onRun
Changed page tracking to onRun instead of onAfterAction.

From the iron:router docs:
"onRun: Called when the route is first run. It is not called again if the route reruns because of a computation invalidation. This makes it a good candidate for things like analytics where you want be sure the hook only runs once. Note that this hook won't run again if the route is reloaded via hot code push."