Fix FlowRouter referrer url
Previously to record the Referrer URL we used context.path which would return FlowRouter's raw path (ex: /posts/:slug/:tag). We want the actual URL. This commit achieves that through FlowRouter._oldExitPath. It would be nice to find a way to do it with a documented/supported FlowRouter key instead of the "_" key.