Editing README
Change the heading to match the project name.
Change the description to match the GitHub description
Add a Table of Contents
Change headings
Add a Maintainers section
Add a link to the CoC
Add the dates and license information to the License section
Add a note about maintainers to contributing section
Add GitHub language blocks to code blocks
Small fixes here and there
Moved some sections for readability.